Added HandlerStyle example
This is an example of the "controllerless action" approach to building your actions.
Each action is defined on its own handler class. All actions have the same name.
A custom UrlPolicy is applied which allows us to achieve the exact same routes as the ControllerAction style even though our actions are defined in a completely different way.